fix: align employee portal identifiers to long

+++ EduOS.Core/DTOs/Portals/EmployeeSelfServiceDtos.cs
@@ -7,8 +7,8 @@ public sealed class EmployeePortalProfileDto
     public string FullName { get; set; } = string.Empty;
     public string Phone { get; set; } = string.Empty;
     public string? Email { get; set; }
-    public int DesignationId { get; set; }
-    public int? DepartmentId { get; set; }
+    public long DesignationId { get; set; }
+    public long? DepartmentId { get; set; }
     public DateTime JoiningDate { get; set; }
     public decimal Salary { get; set; }
     public string? PhotoUrl { get; set; }
